Give the editor its first shortcuts and actions
Save, toggle line comment and close are registered on an "editor" key\ncontext and covered by window tests that really press the keys.\n\nClosing is carried out by the host: a panel that asks the dock to remove\nitself re-enters its own update and panics, and the failure is silent on\nthe real key path.\n\nA failed action now leaves its reason in the status bar instead of doing\nnothing. Not-yet-implemented keys stay unregistered.

+++ crates/editor/src/view/highlight.rs
@@ -198,8 +198,8 @@ fn split_by_line(text: &Rope, source: &str, span: &HighlightSpan) -> Vec<(usize,
         return parts;
     }
 
-    let first_row = text.offset_to_point(start).row as usize;
-    let last_row = text.offset_to_point(end - 1).row as usize;
+    let first_row = text.offset_to_point(start).row;
+    let last_row = text.offset_to_point(end - 1).row;
     for row in first_row..=last_row {
         let line_start = text.line_start_offset(row).min(source.len());
         let line_end = match source[line_start..].find('\n') {
